Reservations_Management/create-reserve/requirements.txt
fastapi
uvicorn
pymysql
sqlalchemy
python-dotenv
psutil